Quadruple DISTVAR
I picked 4096 because it's around the distance of that end slope in Green Hills, which is about the length I imagine a item block "switch" being. I'm being a little conservative though, and it might need to be even higher -- we'll have to play it by ear.

Also changed the types of some of the roulette variables to better match what they are used for.