Allow Rollout Rocks that are being ridden to: * Break bustable FOFs like pushables. * Activate pushable-counting executor activators. * Get pushed by currents/wind.
This takes advantage of the previously-existing hack relating to objects with MF_PUSHABLE infotable (but not active) flags and a nonzero fuse, which is used as the condition to group this under (rather than MT_ROLLOUTROCK specific fuckery).

Thanks to the bug report submitted by our kirbs pal for getting me off my ass this evening.