More logic to spinning your wheels
Use "tire spinning" animations when you're moving on the ground and/or holding accelerate!

The two cases that are changed are:
- If you're in the air and not pressing accelerate, your tires won't spin anymore.
- If you're on the ground, but not moving, but you're still pressing accelerate, they will spin in place. (right now only happens for start boosting, later it will be for spindashing)

Isn't really a tell for anything, just thought it'd be fun attention to detail :)

Additionally, we can use drift frames in the air. Just makes sense because we have air sparks now.