Make sneaker panels push you in spinout AGAIN LOL
The old condition for player movement was...

1. player hasn't finished
2. player not in pain, or player in pain but on the ground

The else condition to this is therefore player has finished or player in pain
and in the air.

I believe K_MomentumToFacing (prior to bots) was only for letting the player
turn around after crossing the finish line, to face the camera.

(K_MomentumToFacing also does nothing in the air.)