Replace the tracking function with Nev3r's version
Was a long time coming :V

Nev3r's supported different FOV. I also expanded it to let it support splitscreen's different screen sizes in a better fashion than the old method of just dividing the final coordinates.

The results are also stored in a struct, instead of modifying passed in variables, so it's also less dunderheaded than what I was doing before overall.