Bot controller linedef
Controls bot behavior while they're touching a tagged sector.

X texture offset is angle to force the bot in (0 is east, 90 is north, 180 is west, 270 is south)

Y texture offset is trick type -- if a bot enters the sector while in trick panel state, then they'll do the input for the trick. Without the effect present, they'll do nothing and eventually tumble.
Y offset 1 does a left trick, 2 is right, 3 is forward, 4 is back/up.