Disable ASM column / span entirely
Causes strange unidentifiable bugs, and testing with tape who has a few framerate issues. perfstats outputted practically the same frames for everyone. Whatever frame drops exist are simply something else and ASM is not helping at all, so decided it's not worth it.

Behind a define anyway, if someone decides to fix it anyway.