Reenable item box respawning
Old code was shit so I removed it (4d67cc632). Turns out
none of that bullshit actually mattered to make this box
respawn and it just needs to go to an invisible state!

Makes use of some nifty flickering code (that was already
there but effectively disabled) shortly before it actually
respawns.