Hyudoro behaves properly without owner set
When it's placed by the level, it's given to whoever has the best position at time of stealing. Since this is meant for duels, this means always the other player.

This behavior is also used as a default whenever its owner gets unset for whatever reason.

If there is literally NO ONE (1P), then rotate in place violently.