HUD tracking polish
Inexplicably, all positional checks were inverted.
This wasn't causing too many problems because the onScreen condition was inverted too, but if it would be above your screen on the opposite side of you, it would be considered onscreen when facing 180 degrees away perfectly.
By swapping viewx/y/z and point->x/y/z, and inverting onScreen, we kill this annoying lack of polish for good.